backend/requirements.txt
# Django core
Django>=5.0,<6.0
Pillow>=10.0.0,<12.0.0


# REST Framework
djangorestframework>=3.14.0,<4.0.0

# CORS for frontend-backend communication
django-cors-headers>=4.0.0,<5.0.0

# Social authentication
dj-rest-auth>=3.0.0,<8.0.0
django-allauth>=0.61.0,<1.0.0

# Required for OAuth token handling
requests>=2.30.0,<3.0.0

# Optional: JWT support (recommended)
djangorestframework-simplejwt>=5.2.0,<6.0.0

# Optional: for managing .env files
python-dotenv>=1.0.0,<2.0.0

# server
gunicorn>=21.2.0,<22.0.0


cryptography>=42.0.0,<43.0.0
requests-oauthlib>=2.0.0,<3.0.0

